refactor(dbt): separate out `DbtCliEventMessage` to `dbt_cli_event.py`
